book_office_project/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
Django==5.1.1
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
minio==7.2.9
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
